+++ example-crs-architecture/secrets.tf
@@ -1,0 +1,17 @@
+resource "kubernetes_secret" "docker-registry" {
+  metadata {
+    name = "ghcr-creds"
+  }
+
+  data = {
+    ".dockerconfigjson" = jsonencode({
+      auths = {
+        "ghcr.io" = {
+          "auth" = var.GHCR_AUTH
+        }
+      }
+    })
+  }
+
+  type = "kubernetes.io/dockerconfigjson"
+}