+++ example-crs-architecture/CHANGELOG.md
@@ -1,5 +1,20 @@
 # CHANGELOG.md
 
+## 2025-01-19
+
+- Updates to outputs.tf
+- Updates to variables.tf
+- Changes hardcoded credentials environment variables
+- Updates to README.md
+
+## 2025-01-15
+
+- Adds example CRS webservice docker image via deployment.tf
+- Adds private load balancer for CRS webservice via deployment.tf
+- Adds docker registry auth via secrets.tf
+- Updates to providers.tf to supprt `kubernetes` provider
+- Updates to README.md
+
 ## 2024-12-03
 
 - Initial commit of example-crs-architecture